Return frozenset instead of dictionary
Previously the distributions were returned as a dictionary mapping the
name of the distribution to the distribution object. This leads to
issues because there can be more than one distribution with the same
name. In that case the distribution that was added last would be the
only one present in the returned dictionary. Returning a frozenset fixes
this problem because distributions are different if they have the same
name but different versions/modules.