Declare user table before creating backend
The record table needs to be declared when the backend gets created
because it can not be declared within a transaction which the unit of
work uses. Declaring the record table requires that the user table is
already declared because the record table depends on the user table.
Therefore we need to declare the user table before creating the backend.